user-story-tests/09-web-component-all-page-styles-low-priority.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Web component all page styles low priority</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>

    <script type="module">
        import { WebComponentAllPageStylesLowPriority } from "./components/WebComponentAllPageStylesLowPriority.js";
        customElements.define(
            "web-component-all-page-styles-low-priority",
            WebComponentAllPageStylesLowPriority
        );
    </script>
    <style>
        @layer resets {
            button {
                border: thick dashed red;
            }
        }
    </style>
    <button>Button outside a shadow tree (thick dashed red)</button>
    <web-component-all-page-styles-low-priority>
    </web-component-all-page-styles-low-priority>

    <script type="module">
        import { adoptStylesFromLayerStatementRule } from "../AdoptStyles.js";

        window.onload = function () {

            const elName = "web-component-all-page-styles-low-priority"

            test(function () {

                const inheritResult = adoptStylesFromLayerStatementRule(document.querySelector(elName))
                assert_true(inheritResult, "inherit found, so returns true");

            }, "Web component all page styles low priority");

        }
    </script>

    <div id="log"></div>

</body>

</html>
